+++ src/NavigatorLazy.tsx
@@ -43,7 +43,7 @@ import WithNavigation from './components/WithNavigation';
 import Registration from './pages/Registration';
 
 import RequestMatchStudent from './pages/student/matching_new/RequestMatch';
-import RequestMatch from './pages/pupil/matching_new/RequestMatch';
+import RequestMatch from './pages/pupil/matching/RequestMatch';
 import Matching from './pages/pupil/Matching';
 import CertificateList from './pages/student/CertificateDetails';
 import NotficationControlPanel from './pages/notification/NotficationControlPanel';
